IPixelStudio/requirements.txt
langchain
langchain_community
langchain_core
Pillow
streamlit
streamlit_option_menu